Renamed wrap layer to perspective layer
